requirements.txt
beautifulsoup4
python-dotenv
playwright
html2text
aiohttp
markdownify
pytest
tqdm
nest-asyncio
python-docx